requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
requests
python-dotenv
sqlalchemy
asyncpg
passlib[argon2]
email-validator
greenlet
